# nodejs/api/swagger.yaml
openapi: 3.0.0
info:
  title: Unified Data Repository Service API file for subscription data
  description: |
    Unified Data Repository Service (subscription data).
    The API version is defined in 3GPP TS 29.504.
    © 2019, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC).
    All rights reserved.
  version: '-'
...
servers:
- url: /
paths:
...
  /subscription-data/subs-to-notify:
    get:
      tags:
      - Subs To Nofify (Collection)
      summary: Retrieves the list of subscriptions
      operationId: QuerySubsToNotify
      parameters:
      - name: ue-id
        in: query
        description: UE id
        required: true
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: supported-features
        in: query
        description: Supported Features
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/SupportedFeatures'
      responses:
        200:
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionDataSubscriptions'
                x-content-type: application/json
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: SubsToNofifyCollection
    post:
      tags:
      - Subs To Nofify (Collection)
      summary: Subscription data subscriptions
      operationId: SubscriptionDataSubscriptions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionDataSubscriptions'
        required: true
      responses:
        201:
          description: Expected response to a valid request
          headers:
            Location:
              description: 'Contains the URI of the newly created resource, according
                to the structure: {apiRoot}/nudr-dr/<apiVersion>/subscription-data/subs-to-notify/{subsId}'
              required: true
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionDataSubscriptions'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      callbacks:
        onDataChange:
          '{request.body#/callbackReference}':
            post:
              requestBody:
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/DataChangeNotify'
                required: true
              responses:
                204:
                  description: Expected response to a valid request
      x-swagger-router-controller: SubsToNofifyCollection
    delete:
      tags:
      - Subs To Notify (Collection)
      summary: Deletes subscriptions identified by a given ue-id parameter
      operationId: RemoveMultipleSubscriptionDataSubscriptions
      parameters:
      - name: ue-id
        in: query
        description: UE ID
        required: true
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: nf-instance-id
        in: query
        description: NF Instance ID
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/NfInstanceId'
      - name: delete-all-nfs
        in: query
        description: Flag to delete subscriptions from all NFs
        required: false
        style: form
        explode: true
        schema:
          type: boolean
      - name: implicit-unsubscribe-indication
        in: query
        description: Implicit Unsubscribe Indication
        required: false
        style: form
        explode: true
        schema:
          type: boolean
      responses:
        204:
          description: Expected response to a successful subscription removal
      x-swagger-router-controller: SubsToNotifyCollection
  /subscription-data/subs-to-notify/{subsId}:
    delete:
      tags:
      - Subs To Notify (Document)
      summary: Deletes a subscriptionDataSubscriptions
      operationId: RemovesubscriptionDataSubscriptions
      parameters:
      - name: subsId
        in: path
        description: Unique ID of the subscription to remove
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        204:
          description: Expected response to a successful subscription removal
      x-swagger-router-controller: SubsToNotifyDocument
    patch:
      tags:
      - Subs To Notify (Document)
      summary: Modify an individual subscriptionDataSubscription
      operationId: ModifysubscriptionDataSubscription
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: subsId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json-patch+json:
            schema:
              minItems: 1
              type: array
              items:
                $ref: '#/components/schemas/PatchItem'
        required: true
      responses:
        204:
          description: Successful response
        403:
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        404:
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        default:
          description: Unexpected error
      x-swagger-router-controller: SubsToNotifyDocument
  /subscription-data/{ueId}/{servingPlmnId}/provisioned-data/trace-data:
    get:
      tags:
      - Trace Data (Document)
      summary: Retrieves the trace configuration data of a UE
      operationId: QueryTraceData
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: servingPlmnId
        in: path
        description: PLMN ID
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarPlmnId'
      - name: If-None-Match
        in: header
        description: Validator for conditional requests, as described in RFC 7232,
          3.2
        required: false
        style: simple
        explode: false
        schema:
          type: string
      - name: If-Modified-Since
        in: header
        description: Validator for conditional requests, as described in RFC 7232,
          3.3
        required: false
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: Expected response to a valid request
          headers:
            Cache-Control:
              description: Cache-Control containing max-age, as described in RFC 7234,
                5.2
              style: simple
              explode: false
              schema:
                type: string
            ETag:
              description: Entity Tag, containing a strong validator, as described
                in RFC 7232, 2.3
              style: simple
              explode: false
              schema:
                type: string
            Last-Modified:
              description: Timestamp for last modification of the resource, as described
                in RFC 7232, 2.2
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceData'
        default:
          description: Generic Error
      x-swagger-router-controller: TraceDataDocument
...
components:
  schemas:
...
    VarPlmnId:
      pattern: ^[0-9]{5,6}$
      type: string
...
    Dnn:
      type: string
...
    SdmSubscription:
      required:
      - callbackReference
      - monitoredResourceUris
      - nfInstanceId
      type: object
      properties:
        nfInstanceId:
          $ref: '#/components/schemas/NfInstanceId'
        implicitUnsubscribe:
          type: boolean
        expires:
          $ref: '#/components/schemas/DateTime'
        callbackReference:
          $ref: '#/components/schemas/Uri'
        amfServiceName:
          $ref: '#/components/schemas/ServiceName'
        monitoredResourceUris:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Uri'
        singleNssai:
          $ref: '#/components/schemas/Snssai'
        dnn:
          $ref: '#/components/schemas/Dnn'
        subscriptionId:
          type: string
        plmnId:
          $ref: '#/components/schemas/PlmnId'
...
    SubscriptionDataSubscriptions:
      required:
      - callbackReference
      - monitoredResourceUris
      type: object
      properties:
        ueId:
          $ref: '#/components/schemas/VarUeId'
        callbackReference:
          $ref: '#/components/schemas/Uri'
        originalCallbackReference:
          $ref: '#/components/schemas/Uri'
        monitoredResourceUris:
          type: array
          items:
            $ref: '#/components/schemas/Uri'
        expiry:
          $ref: '#/components/schemas/DateTime'
        sdmSubscription:
          $ref: '#/components/schemas/SdmSubscription'
        subscriptionId:
          type: string
        supported-features:
          $ref: '#/components/schemas/SupportedFeatures'
    DataChangeNotify:
      type: object
      properties:
        originalCallbackReference:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Uri'
        ueId:
          $ref: '#/components/schemas/VarUeId'
        notifyItems:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/NotifyItem'
        sdmSubscription:
          $ref: '#/components/schemas/SdmSubscription'
        subscriptionDataSubscriptions:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionDataSubscriptions'
...
    VarUeId:
      pattern: ^(imsi-[0-9]{5,15}|nai-.+|msisdn-[0-9]{5,15}|extid-[^@]+@[^@]+|.+)$
      type: string
    SupportedFeatures:
      pattern: ^[A-Fa-f0-9]*$
      type: string
    ProblemDetails:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/Uri'
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          $ref: '#/components/schemas/Uri'
        cause:
          type: string
        invalidParams:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/InvalidParam'
        supportedFeatures:
          $ref: '#/components/schemas/SupportedFeatures'
    Uri:
      type: string
    InvalidParam:
      required:
      - param
      type: object
      properties:
        param:
          type: string
        reason:
          type: string
    PatchItem:
      required:
      - op
      - path
      type: object
      properties:
        op:
          $ref: '#/components/schemas/PatchOperation'
        path:
          type: string
        from:
          type: string
        value:
          nullable: true
    PatchOperation:
      anyOf:
      - type: string
        enum:
        - add
        - copy
        - move
        - remove
        - replace
        - test
      - type: string
...
    NfInstanceId:
      type: string
      format: uuid
    TraceData:
      required:
      - eventList
      - neTypeList
      - traceDepth
      - traceRef
      type: object
      properties:
        traceRef:
          pattern: ^[0-9]{3}[0-9]{2,3}-[A-Fa-f0-9]{6}$
          type: string
        traceDepth:
          $ref: '#/components/schemas/TraceDepth'
        neTypeList:
          pattern: ^[A-Fa-f0-9]+$
          type: string
        eventList:
          pattern: ^[A-Fa-f0-9]+$
          type: string
        collectionEntityIpv4Addr:
          $ref: '#/components/schemas/Ipv4Addr'
        collectionEntityIpv6Addr:
          $ref: '#/components/schemas/Ipv6Addr'
        interfaceList:
          pattern: ^[A-Fa-f0-9]+$
          type: string
      nullable: true
    TraceDepth:
      anyOf:
      - type: string
        enum:
        - MINIMUM
        - MEDIUM
        - MAXIMUM
        - MINIMUM_WO_VENDOR_EXTENSION
        - MEDIUM_WO_VENDOR_EXTENSION
        - MAXIMUM_WO_VENDOR_EXTENSION
      - type: string
    Ipv4Addr:
      pattern: ^(([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])$
      type: string
      example: 198.51.100.1
    Ipv6Addr:
      type: string
      example: 2001:db8:85a3::8a2e:370:7334
      allOf:
      - pattern: ^((:|(0?|([1-9a-f][0-9a-f]{0,3}))):)((0?|([1-9a-f][0-9a-f]{0,3})):){0,6}(:|(0?|([1-9a-f][0-9a-f]{0,3})))$
      - pattern: ^((([^:]+:){7}([^:]+))|((([^:]+:)*[^:]+)?::(([^:]+:)*[^:]+)?))$
...
    Snssai:
      required:
      - sst
      type: object
      properties:
        sst:
          maximum: 255
          minimum: 0
          type: integer
        sd:
          pattern: ^[A-Fa-f0-9]{6}$
          type: string
...
    DateTime:
      type: string
      format: date-time
    NotifyItem:
      required:
      - changes
      - resourceId
      type: object
      properties:
        resourceId:
          $ref: '#/components/schemas/Uri'
        changes:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/ChangeItem'
    ChangeItem:
      required:
      - op
      - path
      type: object
      properties:
        op:
          $ref: '#/components/schemas/ChangeType'
        path:
          type: string
        from:
          type: string
        origValue: {}
        newValue: {}
    ChangeType:
      anyOf:
      - type: string
        enum:
        - ADD
        - MOVE
        - REMOVE
        - REPLACE
      - type: string
...
    ServiceName:
      anyOf:
      - type: string
        enum:
        - nnrf-nfm
        - nnrf-disc
        - nudm-sdm
        - nudm-uecm
        - nudm-ueau
        - nudm-ee
        - nudm-pp
        - namf-comm
        - namf-evts
        - namf-mt
        - namf-loc
        - nsmf-pdusession
        - nsmf-event-exposure
        - nausf-auth
        - nausf-sorprotection
        - nausf-upuprotection
        - nnef-pfdmanagement
        - npcf-am-policy-control
        - npcf-smpolicycontrol
        - npcf-policyauthorization
        - npcf-bdtpolicycontrol
        - npcf-eventexposure
        - npcf-ue-policy-control
        - nsmsf-sms
        - nnssf-nsselection
        - nnssf-nssaiavailability
        - nudr-dr
        - nlmf-loc
        - n5g-eir-eic
        - nbsf-management
        - nchf-spendinglimitcontrol
        - nchf-convergedcharging
        - nnwdaf-eventssubscription
        - nnwdaf-analyticsinfo
      - type: string
...
    PlmnId:
      required:
      - mcc
      - mnc
      type: object
      properties:
        mcc:
          $ref: '#/components/schemas/Mcc'
        mnc:
          $ref: '#/components/schemas/Mnc'
    Mcc:
      pattern: ^\d{3}$
      type: string
    Mnc:
      pattern: ^\d{2,3}$
      type: string
...